requirements.txt
kaggle==1.8.3
langchain-core
langchain-ollama
langgraph
pydantic>=2.0